+++ previous-versions/common/src/codatcommon/models/operations/get_integration.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import integration as shared_integration
+from ...models.shared import integration as shared_integration
 from typing import Optional
 
 
@@ -22,8 +21,6 @@ class GetIntegrationResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
     integration: Optional[shared_integration.Integration] = dataclasses.field(default=None)
     r"""OK"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)