+++ previous-versions/common/src/codatcommon/models/operations/get_integrations_branding.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import branding as shared_branding
-from ..shared import errormessage as shared_errormessage
+from ...models.shared import branding as shared_branding
 from typing import Optional
 
 
@@ -24,8 +23,6 @@ class GetIntegrationsBrandingResponse:
     r"""HTTP response status code for this operation"""
     branding: Optional[shared_branding.Branding] = dataclasses.field(default=None)
     r"""OK"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""