+++ previous-versions/common/src/codatcommon/models/operations/get_webhook.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import webhook as shared_webhook
+from ...models.shared import webhook as shared_webhook
 from typing import Optional
 
 
@@ -22,8 +21,6 @@ class GetWebhookResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""
     webhook: Optional[shared_webhook.Webhook] = dataclasses.field(default=None)