+++ previous-versions/common/src/codatcommon/models/operations/list_rules.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import webhooks as shared_webhooks
+from ...models.shared import webhooks as shared_webhooks
 from typing import Optional
 
 
@@ -28,8 +27,6 @@ class ListRulesResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your `query` parameter was not correctly formed"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""
     webhooks: Optional[shared_webhooks.Webhooks] = dataclasses.field(default=None)