+++ previous-versions/common/src/codatcommon/models/operations/update_profile.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import profile as shared_profile
+from ...models.shared import profile as shared_profile
 from typing import Optional
 
 
@@ -14,8 +13,6 @@ class UpdateProfileResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
     profile: Optional[shared_profile.Profile] = dataclasses.field(default=None)
     r"""OK"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)