+++ previous-versions/common/src/codatcommon/models/shared/apikeys.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 import dataclasses
-from ..shared import apikeydetails as shared_apikeydetails
+from .apikeydetails import APIKeyDetails
 from codatcommon import utils
 from dataclasses_json import Undefined, dataclass_json
 from typing import List, Optional
@@ -11,6 +11,6 @@ from typing import List, Optional
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
 class APIKeys:
-    results: Optional[List[shared_apikeydetails.APIKeyDetails]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('results'), 'exclude': lambda f: f is None }})
+    results: Optional[List[APIKeyDetails]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('results'), 'exclude': lambda f: f is None }})